backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
PyJWT
google-generativeai
wikipedia